requirements.txt
pandas
openpyxl
langchain
langchain-community
langchain-openai
faiss-cpu
python-dotenv
streamlit
rank_bm25
scikit-learn